% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.r
\name{cv,numeric-method}
\alias{cv,numeric-method}
\alias{cv}
\alias{cv,SpatRaster-method}
\title{Coefficient of variation}
\usage{
\S4method{cv}{numeric}(x, na.rm = FALSE)

\S4method{cv}{SpatRaster}(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector or a \code{SpatRaster} "stack."}

\item{na.rm}{Logical: If \code{FALSE} (default), \code{NA} values are propagated.}
}
\value{
Numeric or a \code{SpatRaster}.
}
\description{
Coefficient of variation (standard deviation / mean).
}
\examples{

### numeric
n <- rnorm(100)
cv(n)

# same as
sd(n) / mean(n)

### raster
library(terra)
r1 <- r2 <- r3 <- rast()
nc <- ncell(r1)
r1[] <- runif(nc)
r2[] <- runif(nc)
r3[] <- runif(nc)

r <- c(r1, r2, r3)
sd(r) # standard deviation
cv(r) # coefficient of variation

# same as:
app(r, sd) # standard deviation
app(r, sd) / mean(r) # coefficient of variation
}
