% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenzClim.r
\name{lorenzClim}
\alias{lorenzClim}
\title{Reconstruct future climate variables that match paleoclimate variables}
\usage{
lorenzClim(var, summary, start, end, rcp, gcm, lorenz)
}
\arguments{
\item{var}{Character vector: Name(s) of the variables to return. See \link[=https://www.nature.com/articles/sdata201648/tables/2]{Table 1} and \link[=https://www.nature.com/articles/sdata201648/tables/4]{Table 3} in Lorenz et al. (2016). These can be one or more of the following.  Partial matching is used, and case is ignored.
\itemize{
\item \code{*}: All variables
\item \code{tmax}: Maximum temperature (deg C)
\item \code{tmin}: Maximum temperature (deg C)
\item \code{tmean}: Mean temperature (deg C) (not part of original Lorenz variables)
\item \code{prec}: Precipitation (mm)
\item \code{GDD0} and `GDD5: Growing degree days using base 0 deg C or 5 deg C (days)
\item \code{AET}: Actual evapotranspiration (mm)
\item \code{PET}: Potential evapotranspiration (mm)
\item \code{ETR}: Ratio of AET to PET (unit-less)
\item \code{WDI}: Water deficit index (PET - precipitation; mm)
\item \code{wind}: Wind speed (m / s) (Note: This variable is not available in the paleoclimate set.)
\item \code{VP}: Vapor pressure of water (hPa) (Note: This variable is not available in the paleoclimate set.)
}}

\item{summary}{Character vector: Manner in which to summarize the variable(s). See See \link[=https://www.nature.com/articles/sdata201648/tables/2]{Table 1} in Lorenz et al. (2016). Partial matching is used, and case is ignored:
\itemize{
\item \code{*}: All types (except "\code{raw}").
\item \code{raw}: All monthly values from January of the \code{start} year to December of the \code{end} year.
\item \code{annual}: All annual "summary" and "variability" variables (see next two items).
\item \code{summary}: All annual "summary" variables (sum or mean, depending on the variable).
\item \code{var}: All annual "variability" variables (standard deviation or coefficient of variation, depending on the variable)
\item \code{quarter}: Both quarterly "lower" and "higher" variables (see next two items). Quarters are "calendar" quarters (Q1: January, February, March; Q2: April May, June; and so on). For temperature, ETR, wind, and vapor pressure, monthly values are combined into quarters using averages. For precipitation, GDD0, GDD5, AET, and PET, monthly values are combined into quarters using sums.
\item \code{qlwr}: Values of the quarter with the minimum values of the variables. Returns the "\code{qt_lwr}" rasters (\link[=https://www.nature.com/articles/sdata201648/tables/4]{Table 3}).
\item \code{qhgr}: Values of the quarter with the maximum values of the variables. Returns the "\code{qt_hgr}" rasters (\link[=https://www.nature.com/articles/sdata201648/tables/4]{Table 3}).
\item \code{monthly}: Both monthly "lower" and "higher" variables (see next two items)
\item \code{mlwr}: Minimum value across all 12 months of the year. Returns the "\code{mo_lwr}" rasters (\link[=https://www.nature.com/articles/sdata201648/tables/4]{Table 3}).
\item \code{mhgr}: Maximum value across all 12 months of the year.  Returns the "\code{mo_hgr}" rasters (\link[=https://www.nature.com/articles/sdata201648/tables/4]{Table 3}).
}}

\item{start, end}{Numeric: Start and end years of the period over which the variables are to be calculated. If the period is longer than a year, then the averages across years will be returned. Values must be between 1950 and 2100, inclusive.}

\item{rcp}{Numeric or missing: If only historical (\code{start} and \code{end} are within 1950-2005) variables are desired, then this can be missing. Otherwise, it represents the Representative Concentration Pathway for rasters, and can be any of \code{45}, \code{4.5}, \code{85}, or \code{8.5}.}

\item{gcm}{Character: Name(s) of the general circulation model(s) for which to obtain future variables. If more than one if specified, then the results will be returned as a named list, with one GCM per element in the list. Valid values are listed below. Partial matching is used, and case is ignored.
\itemize{
\item \code{ACCESS1-3}, \code{CanESM2}, \code{CESM1-CAM5}, \code{CNRM-CM5}, \code{CSIRO-Mk3-6-0}, \code{GFDL-CM3}, \code{GISS-E2-R}, \code{HadGEM2-ES}, \code{inmcm4}, \code{IPSL-CM5A-MR}, \code{MIROC5}, \code{MRI-CGCM3}: GCMs named in \link[=https://www.nature.com/articles/sdata201648/tables/4]{Table 3} in Lorenz et al.
\item \code{*}: All GCMs listed above.
}}

\item{lorenz}{Character: Full path of the folder that contains \code{./rcp45} and/or \code{./rcp85}.}
}
\value{
A \code{SpatRaster} with one or more layers. Each layer will be named as per the respetive paleoclimate variable in Lorenz et al. (2016), Table 2.
}
\description{
Lorenz et al. (2016) provide a set of two paleoclimate reconstructions for North America, from 22 or 21 Kybp the the present, plus a set of "historical" reconstructions from 1950 to 2006, plus a third set of CMIP5 "future" climate scenarios for RCPs 4.5 and 8.5. In theory, paleo, historical, and future sets have the same variables. In practice, the rasters for the historical period and future do not always represent the exact same variables as those for the past. For example, the past reconstructions provide the variables \code{an_avg_TMIN}, the annual average of minimum temperature, averaged across a 100 yr time span for each period. For the historical and future scenarios, minimum temperature is represented by one value per month from 1950 to 2100. These functions process the historical and future variables so they are commensurate with those from the paleoclimate reconstructions.

Processing is "seemless" across the historical and future sets. For example, if you desire to calculate temperature across the period 1991 to 2020, the functions will combine the appropriate monthly values from the historical and future sets, so long as both sets are downloaded.

\strong{IMPORTANT}: To process historical and future rasters, they must be downloaded and saved as NetCDF files (the same format in which they are supplied). Data are available on \link[=https://datadryad.org//resource/doi:10.5061/dryad.1597g]{Dryad}
\itemize{
\item Historical files: Historical rasters must be downloaded, unzipped, and saved in a folder named "\code{./historical}". Each subfolder therein should be the same of a GCM, and have the appropriate NetCDF files.
\item Future files: The future files for RCP 4.5 are \emph{must be} place a folder named \code{./rcp45}, and those for RCP 8.5 in \code{./rcp85}. Each of these folders should have one subfolder for each GCM, and the appropriate NetCDF files should be contained inside of them.
}
}
\examples{
\dontrun{
# This example is behind `dontrun{}` because it requires the Lorenz
# rasters to be downloaded and saved to the user's hard drive in the
# appropriately-named folders.

# This is the path where the "/historical", "/rcp45" and "/rcp85" folders
# reside. Each of these needs to have subfolders with each GCM's output. Change
# this according to where you have downloaded and unzipped the files. This
# example is for a Windows system:
lorenz <- 'C:/Ecology/Data/Lorenz et al 2016'

# one variable and one GCM
out <- lorenzClim(
    var = 'tmean',
    summary = '*',
    start = 2005,
    end = 2007,
    rcp = 85,
    gcm = 'ACCESS1-3',
    lorenz = lorenz
)

plot(out)

# two variables and one GCM
out <- lorenzClim(
    var = c('tmean', 'prec'),
    summary = 'annual',
    start = 2005,
    end = 2007,
    rcp = 85,
    gcm = 'ACCESS1-3',
    lorenz = lorenz
)

plot(out)

# two variables and two GCMs
out <- lorenzClim(
    var = c('tmean', 'prec'),
    summary = 'annual',
    start = 2005,
    end = 2007,
    rcp = 85,
    gcm = c('ACCESS1-3', 'CanESM2'),
    lorenz = lorenz
)

out

# one variable, all GCMs, one summary
out <- lorenzClim(
    var = 'PET',
    summary = 'summary',
    start = 2005,
    end = 2007,
    rcp = 85,
    gcm = '*',
    lorenz = lorenz
)

out

# "raw" monthly rasters
out <- lorenzClim(
    var = 'wind',
    summary = 'raw',
    start = 2005,
    end = 2007,
    rcp = 85,
    gcm = 'ACCESS1-3',
    lorenz = lorenz
)

out

# EVERYTHING!!!
out <- lorenzClim(
    var = '*',
    summary = '*',
    start = 2005,
    end = 2007,
    rcp = 85,
    gcm = '*',
    lorenz = lorenz
)

out

}
}
\references{
Lorenz, D.J., Nieto-Lugilde, D., Blois, J.L., Fitzpatrick, M.C., and Williams, J.W.  2016.  Downscaled and debiased climate simulations for North America from 21,000 years ago to 2100AD.  \emph{Scientific Data} 3:160048. \doi{10.1038/sdata.2016.48}

Data for "Downscaled and debiased climate simulations for North America from 21,000 years ago to 2100AD", Dryad \doi{10.5061/dryad.1597g}
}
