% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.r
\name{sd,SpatRaster-method}
\alias{sd,SpatRaster-method}
\alias{sd}
\title{Standard deviation}
\usage{
\S4method{sd}{SpatRaster}(x)
}
\arguments{
\item{x}{A \code{SpatRaster} "stack."}
}
\value{
A \code{SpatRaster}.
}
\description{
Standard deviation.
}
\examples{

### numeric
n <- rnorm(100)
cv(n)

# same as
sd(n) / mean(n)

### raster
library(terra)
r1 <- r2 <- r3 <- rast()
nc <- ncell(r1)
r1[] <- runif(nc)
r2[] <- runif(nc)
r3[] <- runif(nc)

r <- c(r1, r2, r3)
sd(r) # standard deviation
cv(r) # coefficient of variation

# same as:
app(r, sd) # standard deviation
app(r, sd) / mean(r) # coefficient of variation
}
